Move pass ordering functionality to own package
