+++ cpg-core/src/main/java/de/fraunhofer/aisec/cpg/TranslationConfiguration.java
@@ -38,6 +38,9 @@ import de.fraunhofer.aisec.cpg.frontends.LanguageFrontend;
 import de.fraunhofer.aisec.cpg.frontends.cpp.CXXLanguageFrontend;
 import de.fraunhofer.aisec.cpg.frontends.java.JavaLanguageFrontend;
 import de.fraunhofer.aisec.cpg.passes.*;
+import de.fraunhofer.aisec.cpg.passes.order.PassWithDependencies;
+import de.fraunhofer.aisec.cpg.passes.order.PassWithDepsContainer;
+import de.fraunhofer.aisec.cpg.passes.order.RegisterExtraPass;
 import java.io.File;
 import java.lang.reflect.InvocationTargetException;
 import java.util.*;
