+++ cpg-core/src/main/java/de/fraunhofer/aisec/cpg/passes/FilenameMapper.java
@@ -28,6 +28,7 @@ package de.fraunhofer.aisec.cpg.passes;
 import de.fraunhofer.aisec.cpg.TranslationResult;
 import de.fraunhofer.aisec.cpg.graph.Node;
 import de.fraunhofer.aisec.cpg.graph.declarations.TranslationUnitDeclaration;
+import de.fraunhofer.aisec.cpg.passes.order.ExecuteLast;
 import de.fraunhofer.aisec.cpg.processing.IVisitor;
 import de.fraunhofer.aisec.cpg.processing.strategy.Strategy;
 import org.jetbrains.annotations.NotNull;
