+++ cpg-core/src/main/java/de/fraunhofer/aisec/cpg/passes/FunctionPointerCallResolver.kt
@@ -40,6 +40,8 @@ import de.fraunhofer.aisec.cpg.graph.types.IncompleteType
 import de.fraunhofer.aisec.cpg.graph.types.Type
 import de.fraunhofer.aisec.cpg.helpers.IdentitySet
 import de.fraunhofer.aisec.cpg.helpers.SubgraphWalker.ScopedWalker
+import de.fraunhofer.aisec.cpg.passes.order.DependsOn
+import de.fraunhofer.aisec.cpg.passes.order.RequiredFrontend
 import java.util.*
 import java.util.function.Consumer
