+++ cpg-core/src/main/java/de/fraunhofer/aisec/cpg/passes/JavaExternalTypeHierarchyResolver.java
@@ -35,6 +35,8 @@ import de.fraunhofer.aisec.cpg.frontends.java.JavaLanguageFrontend;
 import de.fraunhofer.aisec.cpg.graph.TypeManager;
 import de.fraunhofer.aisec.cpg.graph.types.Type;
 import de.fraunhofer.aisec.cpg.graph.types.TypeParser;
+import de.fraunhofer.aisec.cpg.passes.order.DependsOn;
+import de.fraunhofer.aisec.cpg.passes.order.RequiredFrontend;
 import java.util.HashSet;
 import java.util.List;
 import org.slf4j.Logger;
