+++ cpg-core/src/main/java/de/fraunhofer/aisec/cpg/passes/VariableUsageResolver.java
@@ -39,6 +39,7 @@ import de.fraunhofer.aisec.cpg.graph.statements.expressions.MemberExpression;
 import de.fraunhofer.aisec.cpg.graph.types.*;
 import de.fraunhofer.aisec.cpg.helpers.SubgraphWalker.ScopedWalker;
 import de.fraunhofer.aisec.cpg.helpers.Util;
+import de.fraunhofer.aisec.cpg.passes.order.DependsOn;
 import java.util.*;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
