+++ cpg-core/src/main/java/de/fraunhofer/aisec/cpg/passes/order/RequiredFrontend.kt
@@ -23,19 +23,15 @@
  *                    \______/ \__|       \______/
  *
  */
-package de.fraunhofer.aisec.cpg.passes
+package de.fraunhofer.aisec.cpg.passes.order
 
+import de.fraunhofer.aisec.cpg.frontends.LanguageFrontend
 import kotlin.reflect.KClass
 
 /**
- * Register a dependency for the annotated pass. This ensures that:
- * - the annotated pass is executed after its dependency when `softDependency` is `false`:
- * - the dependency is added to the list of active passes even if not manually specified by the user
- * when `softDependency`is true:
- * - the dependency is not added to the list of active passes - the order is only enforced if the
- * user manually adds the pass
+ * This annotation can only enable a pass for a given [LanguageFrontend]. The pass's accept function
+ * will not be executed when the current language does not match the required langauge.
  */
 @Retention(AnnotationRetention.RUNTIME)
 @Target(AnnotationTarget.CLASS)
-@Repeatable
-annotation class DependsOn(val value: KClass<out Pass>, val softDependency: Boolean = false)
+annotation class RequiredFrontend(val value: KClass<out LanguageFrontend>)
